requirements.txt
temporalio==1.25.0
psycopg[binary]==3.3.3
minio==7.2.20
python-docx==1.2.0
mammoth==1.12.0
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
httpx==0.28.1
